refactor: replace all inputs in auth on ui kit
