+++ @ui/design/settings/colors.scss
@@ -12,6 +12,7 @@ $success-700: #0f9954;
 $success-900: #0a6638;
 $danger-100: #ff8099;
 $danger-300: #f23d61;
+$danger-500: #cc1439;
 $danger-700: #990f2b;
 $danger-900: #660a1d;
 $warning-100: #ffd073;
