+++ @ui/ui-kit/Inputs/Inputs.module.scss
@@ -1,0 +1,99 @@
+@use '../../../@ui/design/settings/colors.scss' as colors;
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap');
+
+.formField {
+  display: flex;
+  align-items: baseline;
+  justify-content: space-between;
+  padding-bottom: 25px;
+}
+
+.input {
+  font-family: 'Inter';
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 24px;
+  width: 100%;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 2px;
+  border: 1px solid colors.$dark-100;
+  color: colors.$light-900;
+  background-color: colors.$dark-500;
+  padding: 5px 11px;
+  margin-top: 4px;
+
+  &:hover {
+    border-color: colors.$light-900;
+  }
+
+  &:focus {
+    border-color: colors.$light-100;
+    color: colors.$light-100;
+    background-color: colors.$dark-300;
+  }
+
+  &:disabled {
+    border-color: colors.$dark-100;
+    color: colors.$dark-100;
+
+    &::placeholder {
+      color: colors.$dark-100;
+    }
+  }
+
+  &:active {
+    border-color: colors.$light-500;
+  }
+
+  &:-webkit-autofill,
+  &:-webkit-autofill:hover, 
+  &:-webkit-autofill:focus {
+    -webkit-text-fill-color: colors.$light-900;
+    -webkit-box-shadow: 0 0 0px 1000px colors.$dark-500 inset;
+  }
+  &:-webkit-autofill:focus {
+    -webkit-text-fill-color: colors.$light-100;
+    -webkit-box-shadow: 0 0 0px 1000px colors.$dark-300 inset;
+  } 
+  
+
+  &::placeholder {
+    color: colors.$light-900;
+  }
+
+  &Error {
+    border: 1px solid colors.$danger-500;
+  }
+}
+
+.errorMessage {
+  color: colors.$danger-500;
+  margin: 4px 0px 0px 5px;
+  float: left;
+  text-align: left;
+  font-size: 13px;
+}
+
+.label {
+  float: left;
+  color: colors.$light-900;
+  font-size: 16px;
+  font-weight: 400;
+
+  &Disabled {
+    color: colors.$dark-100
+  }
+}
+
+.eyeButton {
+  position: absolute;
+  right: 10px;
+  top: 31px;
+  color: white;
+  cursor: pointer;
+  filter: invert(0.9);
+  transform: scale(1.2);
+}
