+++ components/auth/Login/LoginForm.module.scss
@@ -51,32 +51,6 @@
   form {
     padding: 0px 24px;
 
-    label {
-      color: #8d9094;
-      font-size: 14px;
-      line-height: 24px;
-    }
-    input {
-      box-sizing: border-box; /*  ?  */
-      border-radius: 2px;
-      border: 1px solid #4c4c4c;
-      background-color: #171717;
-      width: 100%;
-      color: #8d9094;
-      font-size: 16px;
-      font-family: Inter;
-      padding: 8px 11px;
-      line-height: 24px;
-    }
-    .error {
-      border: 1px solid red;
-    }
-    .errorText {
-      color: red;
-      margin: 4px 0px 0px 5px;
-      text-align: left;
-      font-size: 13px;
-    }
     .button {
       border-radius: 2px;
       background-color: #397df6;
