+++ components/auth/Registration/RegistrationForm.module.scss
@@ -30,7 +30,7 @@
   h1 {
     color: #fff;
     text-align: center;
-    font-family: Inter;
+    font-family: 'Inter';
     font-size: 20px;
     margin-top: 28px;
   }
@@ -45,13 +45,13 @@
   p {
     color: #fff;
     text-align: center;
-    font-family: Inter;
+    font-family: 'Inter';
     margin-top: 20px;
   }
   .SignIn {
     color: #397df6;
     text-align: center;
-    font-family: Inter;
+    font-family: 'Inter';
     font-weight: 600;
     display: block;
     margin: 13px 0 35px;
@@ -61,49 +61,9 @@
     .block {
       margin-top: 22px;
     }
-    label {
-      color: #8d9094;
-      font-size: 14px;
-      font-family: Inter;
-      line-height: 24px;
-    }
-    input {
-      box-sizing: border-box; /*  ?  */
-      border-radius: 2px;
-      border: 1px solid #4c4c4c;
-      background-color: #171717;
-      width: 100%;
-      color: #8d9094;
-      font-size: 16px;
-      font-family: Inter;
-      font-style: normal;
-      font-weight: 400;
-      line-height: 24px;
-      padding: 8px 11px;
-      &:active {
-        border: 1px solid #fff;
-        background: #171717;
-        color: #fff;
-      }
-      &:hover {
-        border: 1px solid #8d9094;
-      }
-      &:disabled {
-        color: #4c4c4c;
-      }
-    }
-    .error {
-      border: 1px solid red;
-    }
-    .errorText {
-      color: red;
-      margin: 4px 0px 0px 5px;
-      text-align: left;
-      font-size: 13px;
-    }
     .button {
       font-size: 16px;
-      font-family: Inter;
+      font-family: 'Inter';
       font-style: normal;
       font-weight: 600;
       line-height: 24px;
