--- components/common/Inputs/Inputs.module.scss
@@ -1,54 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap');
-
-.FormField {
-  display: flex;
-  align-items: baseline;
-  justify-content: space-between;
-  padding-bottom: 25px;
-}
-
-.Input {
-  width: 100%;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 4px;
-  border: 1px solid var(--dark-100, #4c4c4c);
-  font-size: 15px;
-  color: white;
-  background-color: var(--black-a5);
-  box-shadow: 0 0 0 1px var(--black-a9);
-  padding: 8px 11px;
-  line-height: 24px;
-  margin-top: 4px;
-}
-
-.Input:hover {
-  box-shadow: 0 0 0 1px black;
-}
-
-.Input:focus {
-  box-shadow: 0 0 0 2px black;
-}
-
-.Input::selection {
-  background-color: var(--black-a9);
-  color: white;
-}
-
-.Label {
-  float: left;
-  color: #8d9094;
-  font-size: 16px;
-  font-weight: 400;
-}
-
-.EyeButton {
-  position: absolute;
-  right: 10px;
-  bottom: 5px;
-  color: white;
-  cursor: pointer;
-  filter: invert(0.9);
-  transform: scale(1.2);
-}
